requirements.txt
requests
pywebview
QtPy
cefpython3
nuitka
wheel
setuptools